Make HeaderDelimitedMessageHandler write atomic
HeaderDelimitedMessageHandler accepts cancellation token from the call
to JsonRpc.InvokeAsync. If the caller cancels the token at the right
moment, it may happen when the header is already written to the stream,
but the message content is not. If this happens, on the next call the
reciever will get a malformed message - a header from the first message,
a header and content from the next message. This will cause JsonRpc to
error out, close the stream, and cancel the next method.

This fix makes sure that writing in HeaderDelimiteMessageHandler occurs
atomically and cannot be interrupted by cancelled cancellation token.

Fix for #55.