Avoid allocating string to represent whole JSON messages
This commit contains API binary breaking changes, and source breaking changes that will mostly impact any externally defined DelimitedMessageHandler-derived typed (which we don't know of any that exist).

We avoid allocating strings by changing the contract with our message handlers from reading and writing strings to reading and writing JToken. In this implementation, we still allocate contiguous memory for the entire message as an encoded byte buffer (which we were doing previously as well), but we no longer allocate a second "copy" of that for a string representation.

In this commit we define IMessageHandler. This new IMessageHandler interface defines a minimal, and more flexible abstraction  than DelimitedMessageHandler did.

This change also makes use of ValueTask to reduce overhead from Task allocations.