Separate message handlers from formatters
The handlers are concerned with reading and writing messages such that they have separation between each other. The formatters are concerned with actual serialization and deserialization of an individual message.

Formatters and handlers can be (largely) mixed and matched. The one exception to this right now is that the HeaderDelimitedMessageHandler only accepts text-based formatters because it needs to be able to get and set the encoding used as part of the headers.