Keep CancellationToken given to RPC methods working on net472 after method completion
On .NET Framework, CancellationToken becomes half-broken after its CancellationTokenSource is disposed. This doesn't happen on .NET Core, or .NET Framework if the .exe.config file contains this:

```xml
<configuration>
  <runtime>
    <AppContextSwitchOverrides value="Switch.System.Threading.ThrowExceptionIfDisposedCancellationTokenSource=false" />
  </runtime>
</configuration>
```

But as we want to behave properly without that switch on .NET Framework, this change replaces use of a linked CTS with a lone CTS that is temporarily linked by hand such that we can dispose the link later. By doing this, we achieve the same behavior but without a memory leak and without breaking the CancellationToken's use later.