Replace specialized tracker code in formatters with general APIs
Each tracker came with a burden on each formatter that used it that it had to be called at special times and in special ways. This change leverages some recently introduced interfaces so the trackers can take care of most needs themselves, simplifying the job of the formatters.

This includes breaking API changes to the trackers, but we've already made such changes in 2.3, and on the same basis that we aren't aware of any external users of it, it seems like a reasonable opportunity to simplify it.

The IProgress<T> tracker still has one special requirement because of when cleanup happens relative to the $/progress notifications, sadly. When we treat $/progress as just an ordinary method call it can be processed *after* the actual result from the method is received and its callback invoked, which creates a "lossy" progress report because any that are processed on the client after the result callback get dropped. So I ensure $/progress reporting happens directly from the formatter (as it was before) to avoid a significant change in behavior.