Avoid using `AsSpan` on mono
On mono, the `MethodInfo.GetParametres()` method returns a `RuntimeParameterInfo[]` through its `ParameterInfo[]` return type. This works in general due to .NET array covariance. But apparently on mono it breaks when you try to get a slice of the array as a `Span<ParameterInfo>`. We avoid using slice and instead make a copy of the array as a cheap workaround.

Fixes #387