Allow roll-forward policy within the same feature band
I think an improved dev flow is to allow newer SDK versions 
within the same feature band, which is typically a quite safe 
way to guarantee backwards compatibility while gaining 
important patches and minor features automatically if the 
dev box happens to have a newer version (i.e. I have 3.1.301
but it currently cannot be used with the existing policy in 
this repo).