Promote error info into `RemoteRpcException`
Some of this error info was unique to `RemoteInvocationException` but some was replicated across `RemoteMethodNotFoundException` as well. In fact, most RPC error exceptions should include an error code, so promoting these to the base class makes sense.
But not *all* `RemoteRpcException` derived types come from a JSON-RPC error, so we must allow for not having an error code.
We also need to be creative to avoid any backward breaking API changes. So instead of *moving* the properties to the base class, I add them to the base class and 'redefine' them in the derived types to simply forward the value from the base class.