Fix message transmission with args and ETW tracing
In #697 we started using formatter-specific JsonRpcRequest objects on the transmitting side. This causes RPC requests with arguments to fail when ETW tracing is enabled because they try to read arguments using code that was only designed to do so with deserialized messages rather than the one created to be transmitted.

The fix is to avoid trying to deserialize outbound arguments. We could have done an `if` check within the `JsonRpcRequest`-derived classes, but it seemed more complete and safe to just stop reusing classes that were written only with deserialization in mind by creating a far simpler outbound-only request class.