Remove properties inadvertently made public
Back in #908 during the refactoring that created a `FormatterBase` class, properties that used to be explicit interface implementations were made public. These properties are of no interest to the standard user of a formatter so I don't want them exposed publicly. They are still exposed via a public interface so that the advanced cases they were introduced to address are still addressed.

This is not a breaking change relative to our last stable release, because the APIs being removed were added only recently.