Correlate `JoinableTask` tokens across `JsonRpc` instances (by default)
Mutable statics are more or less likely to cause problems with testing or advanced product scenarios. In this case though, it seems *very* likely that every shipping scenario will be drastically simpler by defaulting to sharing state across instances, so that is the default. Advanced cases can opt into isolating those instances however, and I have a test to verify this.

Fixes #983