Fix proxy generation for interfaces A, B when B : A
When multiple interfaces are expressly implemented, and one derives from another, it led to over-subscription of the events on the base interface, causing failures at runtime.
The fix is to ensure we only visit each member once, by visiting each *type* once, regardless of the number of times it is referenced directly or indirectly in the list of types to implement.