Cap the amount of text copied into the stringbuilder in JsonRpcEventSource.Format
Speedometer test shows two LOH allocations occurring in this method when opening a large cs file, when the value passed into the method is a JsonDocument.

1) The ToString call on the document
2) pushing the results of the ToString call into the StringBuilder

It would be non-trivial to get rid of the first allocation, however, the second can be removed by just calling the StringBuilder.AppendOverload that takes in a range from the given string to copy into the StringBuilder.