Use JsonElement.Clone instead of Deserialize
The existing code is using Deserialize to make a copy of the JsonElement in order to hold on to it past the usage of the parent JsonDocument. The problem is that the call to Deserialize isn't passing the JsonSerializerOptions. Thus when this code executes in a trimmed/aot'd app it will fail since reflection is disabled in System.Text.Json.

The fix could be to pass the options into the call, but a better fix all around is to use the Clone method, which is faster than calling Deserialize<JsonElement>.