Avoid 1-2 first chance exceptions on out of order disposal
Disposing of a marshaled object throws when the underlying connection has already been disposed of. These exceptions appear to be innocuous but they are at least false causes for alarm when the program is run under a debugger.

The fix is to arrange to skip when we can predict the exception.