Disable general marshaled objects support for the Nerdbank.MessagePack formatter
It _works_, but it isn't NativeAOT safe. Making this safe in the future may add requirements that may apply to non-NativeAOT users as well, so rather than make it available now and break folks later (or come up with an extra cumbersome API), I think it's best to just leave it unavailable until we can present it in the best possible way.