Eliminate over-build from the `dotnet publish` "test"
The NativeAOTCompatibility.Test project is published as part of our build pipeline as a test. Yes, the publish itself is the test, because in publishing, extra validation is performed by the .NET SDK to ensure that the code is NativeAOT-safe.
But in process of publishing it, we were also *re*-building its dependencies. Not only was this redundant, it was very slow on signed builds. Most importantly, it caused us to ship one set of binaries but archive symbols for an alternate set.

The problem was that we were building and packing first, then we were publishing which re-built but didn't re-pack. As a result, the nupkg packages were stale relative to the pdb's that were on disk from the later build. So we archived symbols based on the loose pdb files, but then we deploy the nupkg binaries.

With this fix, our build gets significantly faster and we only have one set of binaries, so we archive the right symbols.