Replace tedious GenerateShapeFor attributes with method shapes
We now can use method shapes (RPC contracts with `[GenerateShape(Methods)]` on them) to ensure that we have type shapes for all the RPC methods' parameter and return types.

We still need a `Witness` class in a few cases, including just a stub to reach the ShapeProvider to pass to the formatter.

Closes #1269