Enable RPC marshalable objects in `NerdbankMessagePackFormatter`
This was particularly challenging because `[RpcMarshalable]` interfaces can also prescribe 'optional' interfaces which may or may not be implemented by the marshaled object, and this information must be available on the receiving end. We transmit this information (before this PR), but on the receiving end, we would previously emit a dynamic proxy that implements exactly that set of interfaces. To preserve the NativeAOT readiness of `NerdbankMessagePackFormatter` though, we need to use source-generated proxies. A marshalable interface with 8 optional additional interfaces would require 256 proxies to be source generated to have every possible combination, which is ludicrous. Instead, we emit a proxy with just the core interface and another with _all_ the optional interfaces. Receivers who want to test whether an interface is _actually_ available on the remoted object should use the new `Is` and `As<T>` methods, which the proxies can respond to based on intended implementation.

This was also tricky because these optional interfaces _may_ overlap one another in method name or in base types. Since each optional interface has a number prefixed to its method names, we have to be intentional about which prefix to use for a given method that may be reachable from multiple optional interfaces. This unfortunately wasn't thoroughly thought through for dynamic proxies, so for source generated ones we harden it, at the downside of not being _entirely_ compatible with dynamic proxies. But this is a corner case which I suppose no one has hit yet anyway.

I made a design choice to keep the source generated code for proxies very much ignorant as to all the policies around renaming methods. This is because it's not simple, and might evolve over time. But we can't simply evolve all source generated proxies that are compiled into other assemblies, so instead we just have them call into our library for the method transformations at runtime. That has a small runtime cost, but is probably worth retaining the ability to service any bugs we discover later.

- Emit source generated proxies for `[RpcMarshalable]` interfaces _and their optional interfaces_.
- Add new `IClientProxy` interface that can be used to type check proxies that may implement too many interfaces. This is deliberately separated from the pre-existing `IJsonRpcClientProxy` so that RPC isn't relevant, allowing Microsoft.ServiceHub.Framework-generated local proxies to also implement it in a similar way, such that brokered service clients need not know whether RPC is involved to do a proper type check.
- Add several new diagnostic analyzers
- Bump Nerdbank.MessagePack to its latest prerelease version