Switch from PublicAPI files to .NET SDK package validation
This new mechanism will validate a _bunch_ of things about package structure that PublicAPI never touched, including:

1. If an API exists on .NET Standard, it should also exist on all the specific runtime targets that implement that standard.
1. APIs experience no breaking changes across versions.

The only thing that PublicAPI.txt files do that package validation doesn't is a specific file that changes only when public API changes. That can be useful in repos where we have specific PR policies set up to add a required reviewer group to the PR when public APIs are changed, or to make a public API change more obvious. But it's also a real pain to maintain these files, so especially in repos where we _don't_ have such reviewer policies set up, it makes more sense to migrate to the new system.