Drop mutable, idempotent statics from `RpcTargetMetadata`
This means more APIs have `[RequiresDynamicCode]` on them, but these APIs have better alternatives based on PolyType shapes.

The problem with the old design was that an entire process (or AppDomain) would succeed based on any one party calling `RegisterEventArgs<T>()` or `EnableDynamicEventHandlerCreation()`, causing libraries to potentially rely on another party to call those APIs without realizing it, and then malfunctioning when the other code was removed.
With the new design, code will more reliably either succeed or fail on its own merits.

Closes #1275